requirements.txt
numpy==1.16.6
tensorflow-gpu==1.3.0
